sandbox - lift some desktop specific interfaces to common
