+++ src/vs/platform/windows/common/windows.ts
@@ -170,9 +170,37 @@ export interface IPathData {
 	overrideId?: string;
 }
 
+export interface IPathsToWaitFor extends IPathsToWaitForData {
+	paths: IPath[];
+	waitMarkerFileUri: URI;
+}
+
+interface IPathsToWaitForData {
+	paths: IPathData[];
+	waitMarkerFileUri: UriComponents;
+}
+
 export interface IOpenFileRequest {
 	filesToOpenOrCreate?: IPathData[];
 	filesToDiff?: IPathData[];
+}
+
+/**
+ * Additional context for the request on desktop only.
+ */
+export interface IDesktopOpenFileRequest extends IOpenFileRequest {
+	termProgram?: string;
+	filesToWait?: IPathsToWaitForData;
+}
+
+export interface IDesktopRunActionInWindowRequest {
+	id: string;
+	from: 'menu' | 'touchbar' | 'mouse';
+	args?: any[];
+}
+
+export interface IDesktopRunKeybindingInWindowRequest {
+	userSettingsLabel: string;
 }
 
 export interface IWindowConfiguration {
